+++ shoreline.go
@@ -21,10 +21,12 @@ import (
 	"crypto/tls"
 	"encoding/json"
 	"log"
+	"math/rand"
 	"net/http"
 	"os"
 	"os/signal"
 	"syscall"
+	"time"
 
 	"github.com/gorilla/mux"
 
@@ -39,6 +41,7 @@ import (
 )
 
 type (
+	// Config is the Shoreline main configuration
 	Config struct {
 		clients.Config
 		Service disc.ServiceListing `json:"service"`
@@ -54,6 +57,14 @@ const (
 
 func main() {
 	var config Config
+
+	// Init random number generator
+	rand.Seed(time.Now().UnixNano())
+
+	// Set some default config values
+	config.User.MaxFailedLogin = 5
+	config.User.DelayBeforeNextLoginAttempt = 10 // 10 minutes
+	config.User.MaxConcurrentLogin = 100
 
 	if err := common.LoadEnvironmentConfig([]string{"TIDEPOOL_SHORELINE_ENV", "TIDEPOOL_SHORELINE_SERVICE"}, &config); err != nil {
 		log.Panic("Problem loading config", err)